Struct re_web_viewer_server::WebViewerServerHandle

pub struct WebViewerServerHandle {
    pub(crate) local_addr: SocketAddr,
    pub(crate) shutdown_tx: Sender<()>,
}

Sync handle for the WebViewerServer

When dropped, the server will be shut down.

Fields
local_addr: SocketAddrshutdown_tx: Sender<()>
Implementations

impl WebViewerServerHandle

pub fn new( bind_ip: &str, requested_port: WebViewerServerPort ) -> Result<Self, WebViewerServerError>

Create new WebViewerServer to host the Rerun Web Viewer on a specified port. Returns a WebViewerServerHandle that will shutdown the server when dropped.

A port of 0 will let the OS choose a free port.

The caller needs to ensure that there is a tokio runtime running.

pub fn server_url(&self) -> String

Includes http:// prefix